Improve creators overview
Weight wringing or drawing comics rate higher than coloring or
lettering when selecting for the creators "cloud" on the front page.

Also re-implement `creator_contributions` using three underlying views
so it is less complex and can be rematerialized in about 30 ms rather
than 2 seconds.

+++ src/models/mod.rs
@@ -30,17 +30,17 @@ pub trait CloudItem: Ord {
     fn write_item(
         &self,
         out: &mut dyn Write,
-        n: i64,
+        n: i32,
         w: u8,
     ) -> io::Result<()>;
 }
 
 pub struct Cloud<T: CloudItem> {
-    data: Vec<(T, i64, u8)>,
+    data: Vec<(T, i32, u8)>,
 }
 
 impl<T: CloudItem> Cloud<T> {
-    fn from_ordered(data: Vec<(T, i64)>) -> Self {
+    fn from_ordered(data: Vec<(T, i32)>) -> Self {
         let num = data.len();
         let mut data = data
             .into_iter()
